+++ lib/main.dart
@@ -1,4 +1,7 @@
 import 'package:flutter/material.dart';
+import 'package:scanner_bar/components/buttons.dart';
+
+import 'pages/VisualButtons.dart';
 
 void main() {
   runApp(const MyApp());
@@ -15,26 +18,4 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class VisualButtons extends StatelessWidget {
-  const VisualButtons({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Scanner Test'),
-        centerTitle: true,
-      ),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          ElevatedButton(onPressed: () {}, child: Text('QR Code')),
-          ElevatedButton(onPressed: () {}, child: Text('Barcode')),
-          ElevatedButton(onPressed: () {}, child: Text('Photo')),
-        ],
-      ),
-    );
-  }
-}
